moved CurveState under right package
