+++ src/main/java/com/funkydonkies/obstacles/Target.java
@@ -1,7 +1,7 @@
 package com.funkydonkies.obstacles;
 
-import com.funkydonkies.controllers.CurveState;
 import com.funkydonkies.controllers.TargetControl;
+import com.funkydonkies.gamestates.CurveState;
 import com.jme3.bullet.PhysicsSpace;
 import com.jme3.bullet.collision.shapes.BoxCollisionShape;
 import com.jme3.material.Material;
