+++ src/main/java/com/funkydonkies/w4v3/curve/SplineCurve.java
@@ -1,6 +1,6 @@
 package com.funkydonkies.w4v3.curve;
 
-import com.funkydonkies.controllers.CurveState;
+import com.funkydonkies.gamestates.CurveState;
 import com.jme3.bullet.PhysicsSpace;
 import com.jme3.bullet.control.RigidBodyControl;
 import com.jme3.material.Material;
